# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RecipeSearchRequest {
  /** The name of the dish to search for */
  dish: string;
  /** Optional allergy info (e.g. nuts, dairy, gluten) */
  allergies?: string;
  /** Whether the user wants a vegetarian version of the dish */
  vegetarian?: boolean;
}

/**
 * Nutritional information per serving
 */
export type RecipeResultNutrition = {
  /** Calories per serving (e.g. "450 cal") */
  calories: string;
  /** Protein per serving (e.g. "25g") */
  protein: string;
  /** Carbohydrates per serving (e.g. "60g") */
  carbs: string;
  /** Fat per serving (e.g. "15g") */
  fat: string;
};

export interface RecipeResult {
  /** The name of the dish */
  name: string;
  /** Estimated cooking time (e.g. "30 minutes") */
  cookingTime: string;
  /** Number of servings */
  servings: string;
  /** Difficulty level (Easy, Medium, Hard) */
  difficulty: string;
  /** The cuisine type */
  cuisine: string;
  /** A short description of the dish */
  description: string;
  /** List of ingredients with amounts */
  ingredients: string[];
  /** Step-by-step cooking instructions */
  steps: string[];
  /** Optional cooking tips */
  tips: string[];
  /** Whether the recipe is vegetarian */
  isVegetarian: boolean;
  /** Nutritional information per serving */
  nutrition: RecipeResultNutrition;
}

export interface RecipeFeedbackRequest {
  /** The name of the recipe being reviewed */
  recipeName: string;
  /** The name of the person submitting feedback */
  userName: string;
  /**
   * Rating from 1 to 5
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  /** Optional comment or review */
  comment?: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type SubmitRecipeFeedback201 = {
  success: boolean;
};