# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: recipes
    description: Recipe operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /recipes/search:
    post:
      operationId: searchRecipe
      tags: [recipes]
      summary: Search for a recipe using AI
      description: Searches for a recipe by dish name and returns AI-generated recipe data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipeSearchRequest"
      responses:
        "200":
          description: Recipe found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeResult"
        "404":
          description: Recipe not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /recipes/feedback:
    post:
      operationId: submitRecipeFeedback
      tags: [recipes]
      summary: Submit feedback for a recipe
      description: Users can rate and review a recipe with their name and comment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipeFeedbackRequest"
      responses:
        "201":
          description: Feedback submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required:
                  - success
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    RecipeSearchRequest:
      type: object
      properties:
        dish:
          type: string
          description: The name of the dish to search for
        allergies:
          type: string
          description: Optional allergy info (e.g. nuts, dairy, gluten)
        vegetarian:
          type: boolean
          description: Whether the user wants a vegetarian version of the dish
      required:
        - dish

    RecipeResult:
      type: object
      properties:
        name:
          type: string
          description: The name of the dish
        cookingTime:
          type: string
          description: Estimated cooking time (e.g. "30 minutes")
        servings:
          type: string
          description: Number of servings
        difficulty:
          type: string
          description: Difficulty level (Easy, Medium, Hard)
        cuisine:
          type: string
          description: The cuisine type
        description:
          type: string
          description: A short description of the dish
        ingredients:
          type: array
          description: List of ingredients with amounts
          items:
            type: string
        steps:
          type: array
          description: Step-by-step cooking instructions
          items:
            type: string
        tips:
          type: array
          description: Optional cooking tips
          items:
            type: string
        isVegetarian:
          type: boolean
          description: Whether the recipe is vegetarian
        nutrition:
          type: object
          description: Nutritional information per serving
          properties:
            calories:
              type: string
              description: Calories per serving (e.g. "450 cal")
            protein:
              type: string
              description: Protein per serving (e.g. "25g")
            carbs:
              type: string
              description: Carbohydrates per serving (e.g. "60g")
            fat:
              type: string
              description: Fat per serving (e.g. "15g")
          required:
            - calories
            - protein
            - carbs
            - fat
      required:
        - name
        - cookingTime
        - servings
        - difficulty
        - cuisine
        - description
        - ingredients
        - steps
        - tips
        - isVegetarian
        - nutrition

    RecipeFeedbackRequest:
      type: object
      properties:
        recipeName:
          type: string
          description: The name of the recipe being reviewed
        userName:
          type: string
          description: The name of the person submitting feedback
        rating:
          type: integer
          description: Rating from 1 to 5
          minimum: 1
          maximum: 5
        comment:
          type: string
          description: Optional comment or review
      required:
        - recipeName
        - userName
        - rating

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message