# lib/api-zod/src/generated/types/recipeFeedbackRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RecipeFeedbackRequest {
  /** The name of the recipe being reviewed */
  recipeName: string;
  /** The name of the person submitting feedback */
  userName: string;
  /**
   * Rating from 1 to 5
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  /** Optional comment or review */
  comment?: string;
}