# lib/api-zod/src/generated/types/recipeResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RecipeResultNutrition } from "./recipeResultNutrition";

export interface RecipeResult {
  /** The name of the dish */
  name: string;
  /** Estimated cooking time (e.g. "30 minutes") */
  cookingTime: string;
  /** Number of servings */
  servings: string;
  /** Difficulty level (Easy, Medium, Hard) */
  difficulty: string;
  /** The cuisine type */
  cuisine: string;
  /** A short description of the dish */
  description: string;
  /** List of ingredients with amounts */
  ingredients: string[];
  /** Step-by-step cooking instructions */
  steps: string[];
  /** Optional cooking tips */
  tips: string[];
  /** Whether the recipe is vegetarian */
  isVegetarian: boolean;
  /** Nutritional information per serving */
  nutrition: RecipeResultNutrition;
}