# lib/api-zod/src/generated/types/recipeSearchRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RecipeSearchRequest {
  /** The name of the dish to search for */
  dish: string;
  /** Optional allergy info (e.g. nuts, dairy, gluten) */
  allergies?: string;
  /** Whether the user wants a vegetarian version of the dish */
  vegetarian?: boolean;
}